# uws/generated/python-fastapi/src/uws_server/models/uws_version.py
# coding: utf-8

"""
    Universal Worker Service (UWS)

    The Universal Worker Service (UWS) pattern defines how to manage asynchronous execution of jobs on a service.

    The version of the OpenAPI document: 1.2
    Contact: grid@ivoa.net
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class UWSVersion(str, Enum):
    """
    The version of the UWS standard that the server complies with. 
    """

    """
    allowed enum values
    """
    ENUM_1_PERIOD_0 = '1.0'
    ENUM_1_PERIOD_1 = '1.1'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of UWSVersion from a JSON string"""
        return cls(json.loads(json_str))